add service codes

+++ print/Models/DAL/Sender.cs
@@ -11,7 +11,7 @@ namespace print.Models.DAL
     {
         [Required]
         [Description("淘宝获取的卖家user_id")]
-        public int User_Id { get; set; }
+        public long User_Id { get; set; }
         [Required]
         [Description("淘宝获取的卖家nick")]
         public string Sender_Nick { get; set; }
